backend/requirements.txt
spacy
fastapi
uvicorn
supabase
python-jose[cryptography]
python-multipart
scikit-learn
pdfplumber
numpy
